Collider Bug fix
The exception being thrown in the AddListener method of the Collider class seems to be caused by an access violation when the code attempts to read from a null or uninitialized pointer in the listeners array. This can happen if the listener being passed as a parameter is null or if the array has not been properly initialized.